Integer hex nodes (#8)
* Feature: Integer coordinates for all HexNodes

The qrs coordinate system can be modified so that every face (midpoint),
edge (midpoint), and vertex can have integer values for every coordinate
This simplifies rounding operations for nearest edge and nearest vertex,
but adds complexity around excluding integer triples that are not a
vertex, face, or edge. tradeoffs~  ;D <3

First I need to be able to visualize the points I'm describing and their
current and future coordinates

* pretty little markers

next, coordinates of the dots!

* config changes

let's make the codebase a little bit bizarre because that's helpful

* mostly linting changes

* labeled HexNodes

currently labeled as integers. When I'm not tired I'll put both labels
but I would have to do one type of HexNode at a time, because the labels
will be too large otherwise.

* Updating demo

Now scale should allow reading labels.

* ok now they're double-labeled.

* Now we can see both coordinate versions

* Code is communication

The fact that it runs on a machine is largely incidental

* the coordinates only need to show on some grids

* fixed ring subset from when I broke it before